SEQNG-228 Fixes according to PR comments
