+++ app/seqexec-server-gs/src/main/resources/app.conf
@@ -30,6 +30,6 @@ seqexec-engine {
     f2Keywords = true
     gwskeywords = true
     odbQueuePollingInterval = "3 seconds"
-    tops = "tcs=tcs:, ao=ao:, gm=gm:, gc=gc:, gws=wos:, m2=m2:, oiwfs=oiwfs:, ag=ag:, f2=f2:"
+    tops = "tcs=tcs:, ao=ao:, gm=gm:, gc=gc:, gws=ws:, m2=m2:, oiwfs=oiwfs:, ag=ag:, f2=f2:"
     epics_ca_addr_list = "172.17.2.255 172.17.3.255 172.17.102.130 172.17.105.20 172.16.102.130 172.17.106.111 172.17.105.37 172.17.107.50 172.17.55.101 172.17.101.101 172.17.65.255 172.17.102.139 172.17.102.138"
 }
