+++ modules/edu.gemini.seqexec.server/src/main/scala/edu/gemini/seqexec/server/ODBProxy.scala
@@ -1,15 +1,22 @@
 package edu.gemini.seqexec.server
 
 import edu.gemini.pot.sp.SPObservationID
-import edu.gemini.seqexec.odb.SeqExecService
 import edu.gemini.spModel.config2.ConfigSequence
 import edu.gemini.spModel.core.Peer
+import edu.gemini.spModel.core.SPProgramID
+import edu.gemini.seqexec.odb.SeqExecService
+import edu.gemini.seqexec.model.Model.SequenceId
+import edu.gemini.seqexec.server.ConfigUtilOps.ExtractFailure
 import edu.gemini.wdba.session.client.WDBA_XmlRpc_SessionClient
 import edu.gemini.seqexec.engine.Action
 import edu.gemini.seqexec.model.dhs.ImageFileId
 
-import scalaz.{EitherT, \/}
+import scalaz.{Kleisli, EitherT, \/}
 import scalaz.concurrent.Task
+import scala.xml.Elem
+import org.http4s.client.blaze._
+import org.http4s.{Uri, scalaxml}
+import knobs.Config
 
 /**
   * Created by jluhrs on 9/6/16.
@@ -110,10 +117,39 @@ object ODBProxy {
     )
 
     override def queuedSequences(): SeqAction[Seq[SPObservationID]] = EitherT(
-          Task.delay(
-            xmlrpcClient.getObservations(sessionName).toList.map(new SPObservationID(_))
-          ).attempt.map(_.leftMap(e => SeqexecFailure.SeqexecException(e)))
-        )
+      Task.delay(
+        xmlrpcClient.getObservations(sessionName).toList.map(new SPObservationID(_))
+      ).attempt.map(_.leftMap(e => SeqexecFailure.SeqexecException(e)))
+    )
   }
 
 }
+
+case class ODBClientConfig(odbHost: String, port: Int)
+
+case class ODBClient(config: ODBClientConfig) {
+  val httpClient = PooledHttp1Client()
+  // Entity Decoder for xml
+  implicit val decoder = scalaxml.xml()
+
+  def observationTitle(id: SPProgramID, obsId: SequenceId): Task[ExtractFailure \/ String] = {
+    val baseUri = s"http://${config.odbHost}:${config.port}/odbbrowser/observations"
+    val target = Uri.fromString(baseUri).toOption.get +?("programReference", id.stringValue)
+    httpClient.expect[Elem](target).map { xml =>
+      for {
+        x <- xml \\ "observations" \ "observation"
+        if (x \ "id").text == obsId
+        n <- x \ "name"
+      } yield n
+    }.map(_.text).map(\/.right)
+  }
+}
+
+object ODBClient {
+  val DefaultODBBrowserPort: Int = 8442
+  def apply: Kleisli[Task, Config, ODBClient] = Kleisli { cfg: Config =>
+    val odbHost = cfg.require[String]("seqexec-engine.odb")
+    Task.delay(ODBClient(ODBClientConfig(odbHost, DefaultODBBrowserPort)))
+  }
+}
+
